refactor(transform): isolate static value operations
